+++ packages/app/src/systems/CRX/background/services/types.ts
@@ -5,18 +5,29 @@ export type MessageInputs = {
     message: string;
     address: string;
     origin: string;
+    title?: string;
+    favIconUrl?: string;
   };
   sendTransaction: {
     address: string;
     origin: string;
+    title?: string;
+    favIconUrl?: string;
     provider: FuelProviderConfig;
     transaction: string;
   };
   addAsset: {
     asset: Asset;
+  };
+  addAssets: {
+    assets: Asset[];
     origin: string;
+    title?: string;
+    favIconUrl?: string;
   };
   requestConnection: {
     origin: string;
+    title?: string;
+    favIconUrl?: string;
   };
 };