+++ packages/app/src/systems/Core/components/index.tsx
@@ -16,3 +16,5 @@ export * from './VisibilityButton';
 export * from './RouterProvider';
 export * from './AmountVisibility';
 export * from './InputSecurePassword';
+export * from './ConnectInfo';
+export * from './PermissionCard';