+++ packages/app/src/systems/Network/components/index.tsx
@@ -2,4 +2,5 @@ export * from './NetworkDropdown';
 export * from './NetworkForm';
 export * from './NetworkItem';
 export * from './NetworkList';
+export * from './NetworksDialog';
 export * from './NetworkSelector';