+++ packages/app/src/systems/SignUp/components/MnemonicRead/MnemonicRead.tsx
@@ -1,4 +1,5 @@
-import { Stack, Form, Checkbox, Flex, Button, Box } from '@fuel-ui/react';
+import { cssObj } from '@fuel-ui/css';
+import { Stack, Form, Checkbox, Flex, Button, Alert } from '@fuel-ui/react';
 import { useState } from 'react';
 
 import { Header } from '../Header';
@@ -12,7 +13,7 @@ export type MnemonicReadProps = {
 };
 
 export function MnemonicRead({ words, onCancel, onNext }: MnemonicReadProps) {
-  const [isChecked, setChecked] = useState(() => false);
+  const [isSavedChecked, setSavedChecked] = useState(false);
 
   return (
     <Stack gap="$6" align="center">
@@ -23,25 +24,27 @@ export function MnemonicRead({ words, onCancel, onNext }: MnemonicReadProps) {
         alt="Showing your Mnemonic"
       />
       <Header
-        title="Write down your Recovery Phrase"
+        title="Backup your Recovery Phrase"
         subtitle="You will need it on the next step"
       />
-      <Box css={{ width: 400 }}>
+      <Stack css={styles.content} gap="$4">
         <Mnemonic value={words} type="read" />
-      </Box>
-      <Form.Control css={{ flexDirection: 'row' }}>
-        <Checkbox
-          id="confirmSaved"
-          aria-label="Confirm Saved"
-          checked={isChecked}
-          onCheckedChange={(e) => {
-            setChecked(e as boolean);
-          }}
-        />
-        <Form.Label htmlFor="confirmSaved">
-          I saved my passphrase in some secure place
-        </Form.Label>
-      </Form.Control>
+        <Alert status="warning">
+          <Form.Control css={{ flexDirection: 'row' }}>
+            <Checkbox
+              id="confirmSaved"
+              aria-label="Confirm Saved"
+              checked={isSavedChecked}
+              onCheckedChange={(e) => {
+                setSavedChecked(e as boolean);
+              }}
+            />
+            <Form.Label htmlFor="confirmSaved">
+              I have backed up my recovery phrase in a secure place.
+            </Form.Label>
+          </Form.Control>
+        </Alert>
+      </Stack>
       <Flex gap="$4">
         <Button
           color="gray"
@@ -55,7 +58,7 @@ export function MnemonicRead({ words, onCancel, onNext }: MnemonicReadProps) {
           color="accent"
           css={{ width: 130 }}
           onPress={onNext}
-          isDisabled={!isChecked}
+          isDisabled={!isSavedChecked}
         >
           Next
         </Button>
@@ -63,3 +66,19 @@ export function MnemonicRead({ words, onCancel, onNext }: MnemonicReadProps) {
     </Stack>
   );
 }
+
+const styles = {
+  content: cssObj({
+    width: 400,
+
+    '.fuel_alert--icon': {
+      display: 'none',
+    },
+    '.fuel_alert--content': {
+      gap: '$4',
+    },
+    '.fuel_checkbox:focus-within::after': {
+      borderColor: '$yellow5 !important',
+    },
+  }),
+};