+++ packages/app/src/systems/Transaction/components/TxRecipientCard/TxRecipientCard.tsx
@@ -1,7 +1,7 @@
 import { cssObj } from '@fuel-ui/css';
 import { Avatar, Box, Card, Copyable, Flex, Icon, Text } from '@fuel-ui/react';
 import { AddressType } from '@fuel-wallet/types';
-import { isB256, isBech32 } from 'fuels';
+import { Address, isB256, isBech32 } from 'fuels';
 import type { FC } from 'react';
 
 import type { TxRecipientAddress } from '../../types';
@@ -24,8 +24,11 @@ export const TxRecipientCard: TxRecipientCardComponent = ({
   isReceiver,
 }) => {
   const address = recipient?.address || '';
-  const isContract = recipient?.type === AddressType.contract;
   const isValidAddress = isB256(address) || isBech32(address);
+  const fuelAddress = isValidAddress
+    ? Address.fromString(address).toString()
+    : '';
+  const isContract = recipient?.type === AddressType.contract;
 
   return (
     <Card
@@ -43,8 +46,8 @@ export const TxRecipientCard: TxRecipientCardComponent = ({
             <Avatar.Generated
               role="img"
               size="lg"
-              hash={address}
-              aria-label="Generated Address"
+              hash={fuelAddress}
+              aria-label={fuelAddress}
               background="fuel"
             />
           )}