+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @fuel-wallet/sdk
 
+## 0.7.0
+
+### Minor Changes
+
+- [#571](https://github.com/FuelLabs/fuels-wallet/pull/571) [`ac3e822`](https://github.com/FuelLabs/fuels-wallet/commit/ac3e822ec0900e73602b760656f097cdd46e90df) Thanks [@LuizAsFight](https://github.com/LuizAsFight)! - Added asset List Screen
+  Added shortcut in Home Balances to add unknows assets
+  Added DAPP page to "Add Custom Asset"
+  Added fuel.assets()
+  Added fuel.on('assets')
+
+### Patch Changes
+
+- [#593](https://github.com/FuelLabs/fuels-wallet/pull/593) [`f4bd2d5`](https://github.com/FuelLabs/fuels-wallet/commit/f4bd2d563b457de221de46b14cb0cae2ccbe9feb) Thanks [@LuizAsFight](https://github.com/LuizAsFight)! - feat: support multiple assets in `addAsset` method
+
 ## 0.6.0
 
 ### Minor Changes